#1699: split config-mode schema grammar out of ast.go
pkg/config/ast.go was 2021 LOC carrying two distinct concerns: the AST
node types and tree navigation (Node/ConfigTree plus the path-walk and
mutation helpers used by ast_edit.go / ast_groups.go) and the entirely
separate config-mode `set` grammar SSOT (the schemaNode container type,
the large setSchema literal, its groups-wildcard init(), and the
completion / path-resolution functions).

Move the grammar concern out of ast.go into two new same-package files,
finishing the schema_* family that already holds the walker
(schema_walk.go), the validators (schema_validators.go), and the value
classifier (value_type.go):

  - schema.go: the grammar SSOT — schemaNode + isTypedLeaf, the setSchema
    literal, and the groups-wildcard init(). 1347 LOC; it is one
    declarative data tree, kept whole because it is the single source of
    truth that drives completion, flat-set grouping, value-slot help, and
    commit-check validation off one structure (docs/config-schema.md).
  - schema_complete.go: the procedural completion / resolution logic —
    ValueHint, SchemaCompletion, ValueProvider, CompleteSetPath,
    appendTypedValueCompletions, CompleteSetPathWithValues, and
    ResolveConsumedSetPathTokens. 327 LOC; imports only strings.

ast.go drops to 365 LOC (off the refactoring heatmap entirely) and now
holds only the AST data model. keysEqual stays in ast.go: its only
callers are the AST-mutation files (ast_edit.go, ast_groups.go) and no
grammar/completion code references it, so it is an AST-key helper, not a
schema helper.

This is pure intra-package code motion. No symbol was renamed, no logic
edited, and the moved blocks are verbatim. The dual AST shape
(hierarchical `family inet { dhcp; }` vs flat-set `set ... family inet
dhcp`) is untouched because SetPath (ast_edit.go) and walkSchemaNode
(schema_walk.go) are not modified. No exported API change; package-private
symbols resolve across files identically in Go.

Plan reviewed at commit 9379315: Codex PLAN-NEEDS-MINOR, AGY PLAN-READY,
Claude-SMR PLAN-READY — seam unanimously judged real. The
schema_complete.go split is a v2 amendment both Codex and AGY
recommended so neither new file lands on the audit heatmap.

Validation: gofmt clean; go build ./... and go vet ./pkg/config/...
clean; go test ./pkg/config/... passes; the completion + grammar-grouping
+ schema-walk tests pass 5/5 in a flake loop. The two failing tests in
pkg/dataplane/userspace are a pre-existing sandbox unix-socket bind
limitation that reproduces identically on clean master and is unrelated
to this config-only change.
